Add custom `changepassword` command to replace django built-in
This commit replaces Django's `manage.py changepassword` functionality.
This change is necessary because we hash the passwords client-side in the
Admin/Registration/Login pages using javascript and then authenticate
against that on the server. However, when using the commandline changepassword
the first hash does not happen which causes future authentications through
the web interface to fail