Add adjacency calculator to set of calculators
Adjacency calculator calculates average cost/travel/etc.
between all base geounits for a district and plan. It
assigns a score to a district that is the average adjacency
between all base geounits of a district or a normalized
score for an entire region/plan based on the individual
district scores and the average adjacency between all
base geounits for a region.

To speed up calculations and queries an in-memory key-value
store (Redis) is used to store adjacency data.