Moved this to end of mapinit - only seems to be required for new maps
This event listener/trigger was causing some buggy behavior where on
changes in zoom and thematic map for districts were canceling each
other out and removing choropleths.

This seems to be required only on map initialization, so this just gets
called at the end of the mapinit function, once instead.