Use localhost instead of site.domain when map server not defined
Fixes an issue where shapefile downloads could not be generated
because celery/django could not connect to the map server b/c
the url was incorrect.

I think this is a less brittle approach than using the site domain.
Site domain is set to example.com by default in django, and is
almost never set otherwise. This assumes that the map server is
running locally. In most cases this is true, unless it is on a
different server, which in that case it will be set in the XML
file and this script will never get to this if statement.